Move `ConfigurationCacheDegradationController` to internal API package
